[cudamapper] Moved print_paf to cudamapper_utils
