+++ cudamapper/src/cudamapper_utils.hpp
@@ -10,13 +10,19 @@
 
 #pragma once
 
-#include <algorithm>
+#include <mutex>
 #include <vector>
 
 #include <claragenomics/cudamapper/types.hpp>
 
 namespace claragenomics
 {
+
+namespace io
+{
+class FastaParser;
+}; // namespace io
+
 namespace cudamapper
 {
 /// \brief given a vector of overlaps, combines all overlaps from the same read pair
@@ -48,6 +54,20 @@ namespace cudamapper
 /// \param fused_overlaps Output vector for fused overlaps
 /// \param unfused_overlaps vector of overlaps, sorted by (query_id, target_id) combination and query_start_position
 void fuse_overlaps(std::vector<Overlap>& fused_overlaps, const std::vector<Overlap>& unfused_overlaps);
+
+/// \brief prints overlaps to stdout in <a href="https://github.com/lh3/miniasm/blob/master/PAF.md">PAF format</a>
+/// \param overlaps vector of overlap objects
+/// \param cigar cigar strings
+/// \param query_parser needed for read names and lenghts
+/// \param target_parser needed for read names and lenghts
+/// \param kmer_size minimizer kmer size
+/// \param write_output_mutex mutex that enables exclusive access to output stream
+void print_paf(const std::vector<Overlap>& overlaps,
+               const std::vector<std::string>& cigar,
+               const io::FastaParser& query_parser,
+               const io::FastaParser& target_parser,
+               std::int32_t kmer_size,
+               std::mutex& write_output_mutex);
 
 } // namespace cudamapper
 } // namespace claragenomics
