+++ cudamapper/src/overlapper_triggered.cu
@@ -8,15 +8,15 @@
  * license agreement from NVIDIA CORPORATION is strictly prohibited.
  */
 
-#include <cub/cub.cuh>
+#include "overlapper_triggered.hpp"
 
 #include <fstream>
 #include <cstdlib>
 
-#include <claragenomics/utils/cudautils.hpp>
+#include <cub/cub.cuh>
+#include <thrust/execution_policy.h>
 
-#include "cudamapper_utils.hpp"
-#include "overlapper_triggered.hpp"
+#include <claragenomics/utils/cudautils.hpp>
 
 #ifndef NDEBUG // only needed to check if input is sorted in assert
 #include <algorithm>
